+++ src/faucet/views.rs
@@ -45,11 +45,15 @@ pub fn Faucet(target_network: Network) -> impl IntoView {
     let (drip_amount, faucet_tx_base_url) = match target_network {
         Network::Mainnet => (
             crate::constants::MAINNET_DRIP_AMOUNT.clone(),
-            option_env!("FAUCET_TX_URL_MAINNET").and_then(|url| Url::parse(url).ok()),
+            RwSignal::new(
+                option_env!("FAUCET_TX_URL_MAINNET").and_then(|url| Url::parse(url).ok()),
+            ),
         ),
         Network::Testnet => (
             crate::constants::CALIBNET_DRIP_AMOUNT.clone(),
-            option_env!("FAUCET_TX_URL_CALIBNET").and_then(|url| Url::parse(url).ok()),
+            RwSignal::new(
+                option_env!("FAUCET_TX_URL_CALIBNET").and_then(|url| Url::parse(url).ok()),
+            ),
         ),
     };
     let topup_req_url = option_env!("FAUCET_TOPUP_REQ_URL");
@@ -196,7 +200,7 @@ pub fn Faucet(target_network: Network) -> impl IntoView {
                                     .into_iter()
                                     .map(|(msg, sent)| {
                                         let (cid, status) = if sent {
-                                            let cid = faucet_tx_base_url
+                                            let cid = faucet_tx_base_url.get()
                                                 .as_ref()
                                                 .and_then(|base_url| format_url(base_url, SearchPath::Transaction ,&msg.to_string()).ok())
                                                 .map(|tx_url| {
@@ -229,10 +233,30 @@ pub fn Faucet(target_network: Network) -> impl IntoView {
                 }
             }}
         </div>
-        <div class="flex flex-col items-center">
-            <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-2 rounded-full">
-              <a href="/faucet">Back to faucet list</a>
-            </button>
+        <div class="flex justify-center space-x-4">
+        {move || {
+            match faucet_tx_base_url.get() {
+                Some(ref base_url) => match format_url(base_url, SearchPath::Address, &faucet.get().get_sender_address()) {
+                    Ok(addr_url) => view! {
+                        <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-2 rounded-full">
+                            <a
+                                href={addr_url.to_string()}
+                                target="_blank"
+                                rel="noopener noreferrer"
+                            >
+                                "Transaction History"
+                            </a>
+                        </button>
+                    }
+                    .into_any(),
+                    Err(_) => ().into_any(),
+                },
+                None => ().into_any(),
+            }
+        }}
+        <button class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-1 px-2 rounded-full">
+            <a href="/faucet">Back to faucet list</a>
+        </button>
         </div>
     }
 }
@@ -253,11 +277,8 @@ pub fn Faucets() -> impl IntoView {
 #[component]
 pub fn Faucet_Calibnet() -> impl IntoView {
     let rpc_context = RpcContext::use_context();
-
-    // update the rpc context to the calibnet url
-    Effect::new(move |_| {
-        rpc_context.set(Provider::get_network_url(Network::Testnet));
-    });
+    // Set rpc context to calibnet url
+    rpc_context.set(Provider::get_network_url(Network::Testnet));
 
     view! {
         <Title text="Filecoin Faucet - Calibration Network" />
@@ -275,11 +296,8 @@ pub fn Faucet_Calibnet() -> impl IntoView {
 #[component]
 pub fn Faucet_Mainnet() -> impl IntoView {
     let rpc_context = RpcContext::use_context();
-
-    // update the rpc context to the mainnet url
-    Effect::new(move |_| {
-        rpc_context.set(Provider::get_network_url(Network::Mainnet));
-    });
+    // Set rpc context to mainnet url
+    rpc_context.set(Provider::get_network_url(Network::Mainnet));
 
     view! {
         <Title text="Filecoin Faucet - Mainnet" />